# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DMS - Document Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /protocols:
    get:
      operationId: listProtocols
      tags: [protocols]
      summary: List protocols
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
        - name: dossierId
          in: query
          schema:
            type: ["integer", "null"]
        - name: assignedToMe
          in: query
          schema:
            type: ["boolean", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Protocol list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProtocolList"
    post:
      operationId: createProtocol
      tags: [protocols]
      summary: Create and register a new protocol
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProtocolInput"
      responses:
        "201":
          description: Created protocol
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Protocol"
  /protocols/{id}:
    get:
      operationId: getProtocol
      tags: [protocols]
      summary: Get protocol by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Protocol
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Protocol"
        "404":
          description: Not found
    patch:
      operationId: updateProtocol
      tags: [protocols]
      summary: Update protocol metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProtocolUpdate"
      responses:
        "200":
          description: Updated protocol
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Protocol"
  /protocols/{id}/cancel:
    post:
      operationId: cancelProtocol
      tags: [protocols]
      summary: Cancel (annul) a protocol
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelProtocolInput"
      responses:
        "200":
          description: Cancelled protocol
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Protocol"
...
  /signatures:
    get:
      operationId: listSignatures
      tags: [signatures]
      summary: List signature requests
      parameters:
        - name: documentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
        - name: pendingForMe
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Signature requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SignatureRequest"
    post:
      operationId: createSignatureRequest
      tags: [signatures]
      summary: Request signatures on a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignatureRequestInput"
      responses:
        "201":
          description: Created signature request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignatureRequest"
  /signatures/{id}/sign:
    post:
      operationId: signDocument
      tags: [signatures]
      summary: Sign or reject a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignatureAction"
      responses:
        "200":
          description: Updated signature request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignatureRequest"

  # --- USERS ---
...
components:
  schemas:
...
    ProtocolList:
      type: object
      required:
        - items
        - total
        - page
        - limit
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Protocol"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    Protocol:
      type: object
      required:
        - id
        - number
        - year
        - type
        - status
        - subject
        - registeredAt
        - registeredById
        - registeredByName
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        number:
          type: string
        year:
          type: integer
        type:
          type: string
        status:
          type: string
        subject:
          type: string
        description:
          type: ["string", "null"]
        sender:
          type: ["string", "null"]
        recipients:
          type: array
          items:
            type: string
        ccRecipients:
          type: array
          items:
            type: string
        channel:
          type: ["string", "null"]
        confidentiality:
          type: string
          default: normal
        priority:
          type: string
          default: normal
        dossierId:
          type: ["integer", "null"]
        dossierTitle:
          type: ["string", "null"]
        classificationId:
          type: ["integer", "null"]
        classificationCode:
          type: ["string", "null"]
        documentId:
          type: ["integer", "null"]
        assignedToId:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        registeredById:
          type: integer
        registeredByName:
          type: string
        cancelledAt:
          type: ["string", "null"]
        cancelReason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        driveFolder:
          type: ["string", "null"]
        registeredAt:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    ProtocolInput:
      type: object
      required:
        - type
        - subject
      properties:
        type:
          type: string
        subject:
          type: string
        description:
          type: string
        sender:
          type: string
        recipients:
          type: array
          items:
            type: string
        ccRecipients:
          type: array
          items:
            type: string
        channel:
          type: string
        confidentiality:
          type: string
        priority:
          type: string
        dossierId:
          type: ["integer", "null"]
        classificationId:
          type: ["integer", "null"]
        documentId:
          type: ["integer", "null"]
        assignedToId:
          type: ["integer", "null"]
        notes:
          type: string
    ProtocolUpdate:
      type: object
      properties:
        subject:
          type: string
        description:
          type: string
        sender:
          type: string
        recipients:
          type: array
          items:
            type: string
        priority:
          type: string
        confidentiality:
          type: string
        dossierId:
          type: ["integer", "null"]
        classificationId:
          type: ["integer", "null"]
        assignedToId:
          type: ["integer", "null"]
        notes:
          type: string
    CancelProtocolInput:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
          minLength: 1
...
    SignatureRequest:
      type: object
      required:
        - id
        - documentId
        - status
        - signatories
        - createdAt
      properties:
        id:
          type: integer
        documentId:
          type: integer
        documentTitle:
          type: ["string", "null"]
        status:
          type: string
        type:
          type: string
          default: internal
        signatories:
          type: array
          items:
            $ref: "#/components/schemas/Signatory"
        requestedById:
          type: integer
        requestedByName:
          type: string
        note:
          type: ["string", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
    Signatory:
      type: object
      required:
        - userId
        - userName
        - order
        - status
      properties:
        userId:
          type: integer
        userName:
          type: string
        order:
          type: integer
        status:
          type: string
        signedAt:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
    SignatureRequestInput:
      type: object
      required:
        - documentId
        - signatories
      properties:
        documentId:
          type: integer
        type:
          type: string
          default: internal
        signatories:
          type: array
          items:
            type: object
            required:
              - userId
              - order
            properties:
              userId:
                type: integer
              order:
                type: integer
        note:
          type: string
        expiresAt:
          type: string
    SignatureAction:
      type: object
      required:
        - action
      properties:
        action:
          type: string
        note:
          type: string

    # --- USERS ---
...