# lib/api-zod/src/generated/types/protocolInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DMS - Document Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface ProtocolInput {
  type: string;
  subject: string;
  description?: string;
  sender?: string;
  recipients?: string[];
  ccRecipients?: string[];
  channel?: string;
  confidentiality?: string;
  priority?: string;
  /** @nullable */
  dossierId?: number | null;
  /** @nullable */
  classificationId?: number | null;
  /** @nullable */
  documentId?: number | null;
  /** @nullable */
  assignedToId?: number | null;
  notes?: string;
}