# lib/api-zod/src/generated/types/protocolList.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DMS - Document Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { Protocol } from './protocol';

export interface ProtocolList {
  items: Protocol[];
  total: number;
  page: number;
  limit: number;
}