# lib/api-zod/src/generated/types/signatory.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DMS - Document Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Signatory {
  userId: number;
  userName: string;
  order: number;
  status: string;
  /** @nullable */
  signedAt?: string | null;
  /** @nullable */
  note?: string | null;
}