# lib/api-zod/src/generated/types/signatureRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DMS - Document Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { Signatory } from './signatory';

export interface SignatureRequest {
  id: number;
  documentId: number;
  /** @nullable */
  documentTitle?: string | null;
  status: string;
  type?: string;
  signatories: Signatory[];
  requestedById?: number;
  requestedByName?: string;
  /** @nullable */
  note?: string | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  expiresAt?: string | null;
}